Introduce SXE_CAST() and SXE_CAST_NOCONST() macros.
Use these macros to cast a pointer to uintptr_t, then to your target type.
This replaces various idioms that do not work on 64-bit platforms:

   (char *)(long)"hello"
   (SXE *)(unsigned)expression

Instead, you would now say something like this:

   SXE_CAST_NOCONST(char *, "hello")
   SXE_CAST(SXE *, expression)

The two macros are identical, but the _NOCONST variant makes it more obvious
that the intent is to lose the 'const'.

The new macros are in lib-sxe-util, so the following packages do not use them,
because they are built before lib-sxe-util:

   lib-tap
   lib-port
   lib-mock
   lib-sxe-log