Always ev_{io,async}_init() if sxe->socket is set.
Without this change, if you call sxe_close() on an SXE that has not
initialized its 'async' object, ev_async_stop() could access invalid memory.
For some reason it didn't *always* do this.

The same goes for ev_io_stop if we hadn't called ev_io_init(); but that was
already protected by checking whether the socket was valid. We now use the
same flag for both ev watchers.