sxe_send_buffers(): sxe_send() a list of buffers
Introduce a writev()-like construct for sending a list of buffers in a single
call, and notifying when the whole list has been sent. This is extremely
useful when the caller needs to perform a sequence of send() operations:
rather than having to keep track of how much has been sent in the application,
sxe_send_buffers() will automatically send all the data, advancing through the
list until the data is complete.

The buffer list itself is managed by the caller, and the memory it points to
must not be changed until sxe_send_buffers() invokes the on_complete callback.